leave type

+++ Onyx/Models/StoredProcedure/CompanyLeave_GetRow_Result.cs
@@ -12,12 +12,12 @@
         public DateTime? EditDt { get; set; }
         public string CoCd { get; set; }
         public string LvCd { get; set; }
-        public decimal LvMax { get; set; }
+        public int LvMax { get; set; }
         public string Accrued { get; set; }
         public string EnCash { get; set; }
-        public decimal EnCashMinLmt { get; set; }
+        public int EnCashMinLmt { get; set; }
         public decimal PayFact { get; set; }
-        public decimal AccrLmt { get; set; }
+        public int AccrLmt { get; set; }
         public string ServicePrd { get; set; }
     }
 }
